feat: exit if file exists

+++ export_module.py
@@ -2,6 +2,11 @@
 import pandas as pd
 import json
 import re
+from sys import exit
+from os.path import exists
+
+#messages
+error_file_exists = 'ERROR: file with same pathname allready exists.'
 
 #helper functions
 
@@ -15,16 +20,22 @@ def rmv_file_extension(pathname):
 
 #export functions
 
-def write_to_jason(json_data, pathname):
+def write_to_jason(json_data, pathname, overwrite = False):
+    if(overwrite is False and exists(pathname)): exit(error_file_exists)
+
     file = open(file=pathname,mode='w')
     file.write(json.dumps(obj=json_data,indent=2))
     file.close()
 
-def write_to_csv(json_data, pathname):
+def write_to_csv(json_data, pathname, overwrite = False):
+    if(overwrite is False and exists(pathname)): exit(error_file_exists)
+
     df = pd.DataFrame(json_data)
     df.to_csv(path_or_buf=pathname,index=False,header=False,mode='w')
 
-def write_to_excel(json_data, pathname): #write to file with xlsx excel format
+def write_to_excel(json_data, pathname, overwrite = False): #write to file with xlsx excel format
+    if(overwrite is False and exists(pathname)): exit(error_file_exists)
+
     df = pd.DataFrame(json_data)
     df.to_excel(excel_writer=pathname,index=False,header=False)
